Do not add milestone to backport PRs
The milestone is decided by the original PR and often does not match
the required milestone for a backport PR to produce a correct
changelog entry.

By not automatically setting the milestone, maintainers are required
to consider and set it themselves. It is currently believe that edge
cases around target branch to milestone mapping (such as out of date
milestones) mean that this human step is less error prone than code
would be.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>