Match any reasonable patch version
Although semver requires that every bump of a major or minor version resets the patch version,
it is not always the case that the tag will exist in the repository.
For example, the first release might be made privately in a fork and might never be made public
because of bugs or security concerns.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>