Automatically resolve docs only conflicts by keeping "our" changes
Technical writers have explained that most of the time spent backporting is consumed by manually performing the cherry-pick due to merge conflicts.
In each case, their workflow is almost always to manually resolve the conflict by taking the changes from the source PR (--ours).

This change intends to automatically perform this action, but *only* for PRs that exclusively touch docs.

Technical writers will then be required to scrutinize the PR patch and not the conflict markers.
In nearly all cases, it is expected that the PR patch will be acceptable.
In the case that it is not, technical writers will be expected to update the backport PR appropriately.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>